Mailmind/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.4
djangorestframework==3.15.2
idna==3.10
pillow==11.0.0
python-dotenv==1.0.1
requests==2.32.3
soupsieve==2.6
sqlparse==0.5.3
urllib3==2.2.3
